refactor!: plotting backend is now optional
