+++ setup.cfg
@@ -39,32 +39,12 @@ zip_safe = False
 install_requires =
   numpy >=1.13.3
   boost-histogram ~=0.11.0
-  matplotlib >=3.0
-  scipy >=1.4
-  uncertainties >=3
   histoprint >=1.4
-  mplhep >=0.2.2
   typing_extensions; python_version<"3.8"
 
 [options.entry_points]
 console_scripts=
   hist=hist.classichist:main
-
-[options.extras_require]
-dev =
-  pytest >=4.6
-  ipykernel
-test =
-  pytest >=4.6
-docs =
-  nbsphinx
-  recommonmark >=0.5.0
-  Sphinx >=3.0.0
-  sphinx_copybutton
-  sphinx_rtd_theme >=0.5.0
-  ipython
-plot =
-  matplotlib >=3.0
 
 [options.packages.find]
 where = src
